refactor: rename useRequestsFilter filename
